Crate problem043

Problem 43 (Sub-string divisibility)

Solution detail

The easiest way to solve this problem is by using a depth-first search, building up the number one digit at a time, and pruning branches whenever one of the divisibility conditions is not satisfied.

In other words, we consider all those choices for the first four digits which are divisible by 2, then for each of those, consider all choices for the next digit which meets the condition for divisibility by 3, and so on until the whole number has been built.

Structs

SubstringTree
The information that is held about the current state during the tree search.

SubstringTreeStep
A description of a step that can be taken in the search tree.

Functions

answer
Solve the problem, returning the answer as a String

main

solve
Find the sum of all 10-digit pandigital numbers satisfying the divisibility conditions.